split tagtransform into two classes and an interface
