+++ tagtransform.hpp
@@ -16,41 +16,75 @@ extern "C" {
 }
 #endif
 
-
-
-class tagtransform {
+class tagtransform_t
+{
 public:
-	tagtransform(const options_t *options_);
-	~tagtransform();
+    static std::unique_ptr<tagtransform_t>
+    make_tagtransform(options_t const *options);
 
-    bool filter_tags(osmium::OSMObject const &o,
-                     int *polygon, int *roads, const export_list &exlist,
-                     taglist_t &out_tags, bool strict = false);
-    unsigned filter_rel_member_tags(const taglist_t &rel_tags,
-        const multitaglist_t &member_tags, const rolelist_t &member_roles,
-        int *member_superseded, int *make_boundary, int *make_polygon, int *roads,
-        const export_list &exlist, taglist_t &out_tags, bool allow_typeless = false);
+    virtual ~tagtransform_t() = 0;
+
+    virtual bool filter_tags(osmium::OSMObject const &o, int *polygon,
+                             int *roads, export_list const &exlist,
+                             taglist_t &out_tags, bool strict = false) = 0;
+
+    virtual unsigned filter_rel_member_tags(
+        taglist_t const &rel_tags, multitaglist_t const &member_tags,
+        rolelist_t const &member_roles, int *member_superseded,
+        int *make_boundary, int *make_polygon, int *roads,
+        export_list const &exlist, taglist_t &out_tags,
+        bool allow_typeless = false) = 0;
+};
+
+class c_tagtransform_t : public tagtransform_t
+{
+public:
+    c_tagtransform_t(options_t const *options);
+
+    bool filter_tags(osmium::OSMObject const &o, int *polygon, int *roads,
+                     export_list const &exlist, taglist_t &out_tags,
+                     bool strict = false) override;
+
+    unsigned filter_rel_member_tags(taglist_t const &rel_tags,
+                                    multitaglist_t const &member_tags,
+                                    rolelist_t const &member_roles,
+                                    int *member_superseded, int *make_boundary,
+                                    int *make_polygon, int *roads,
+                                    export_list const &exlist,
+                                    taglist_t &out_tags,
+                                    bool allow_typeless = false) override;
 
 private:
-    bool lua_filter_basic_tags(osmium::OSMObject const &o,
-                               int *polygon, int *roads, taglist_t &out_tags);
-    bool c_filter_basic_tags(osmium::OSMObject const &o, int *polygon,
-                             int *roads, const export_list &exlist,
-                             taglist_t &out_tags, bool strict);
-    unsigned int lua_filter_rel_member_tags(const taglist_t &rel_tags,
-        const multitaglist_t &members_tags, const rolelist_t &member_roles,
-        int *member_superseded, int *make_boundary, int *make_polygon, int *roads,
-        taglist_t &out_tags);
-    void check_lua_function_exists(const std::string &func_name);
-
-
-	const options_t* options;
-	const bool transform_method;
-#ifdef HAVE_LUA
-	lua_State *L;
-    const std::string m_node_func, m_way_func, m_rel_func, m_rel_mem_func;
-#endif
-
+    options_t const *m_options;
 };
+
+#ifdef HAVE_LUA
+class lua_tagtransform_t : public tagtransform_t
+{
+public:
+    lua_tagtransform_t(options_t const *options);
+    ~lua_tagtransform_t();
+
+    bool filter_tags(osmium::OSMObject const &o, int *polygon, int *roads,
+                     export_list const &exlist, taglist_t &out_tags,
+                     bool strict = false) override;
+
+    unsigned filter_rel_member_tags(taglist_t const &rel_tags,
+                                    multitaglist_t const &member_tags,
+                                    rolelist_t const &member_roles,
+                                    int *member_superseded, int *make_boundary,
+                                    int *make_polygon, int *roads,
+                                    export_list const &exlist,
+                                    taglist_t &out_tags,
+                                    bool allow_typeless = false) override;
+
+private:
+    void check_lua_function_exists(std::string const &func_name);
+
+    options_t const *m_options;
+    lua_State *L;
+    std::string m_node_func, m_way_func, m_rel_func, m_rel_mem_func;
+};
+#endif
 
 #endif //TAGTRANSFORM_H
